+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.2.0
+
+### Improvements
+
+- Add an extra generic param to allow you to omit certain props from the base elements props
+
+### Bug Fixes
+
+- Omit key and ref from the combined type in `ComponentProps`
+- Mark the package as side effect free
+
 ## 1.1.2
 
 ### Bug Fixes